Define RPMTAG_NOT_FOUND in uint32_t terms
The bit-level value doesn't change, but this makes the value
explicitly unsigned for the compiler, which is what we care about here.

+++ include/rpm/rpmtag.h
@@ -389,7 +389,7 @@ typedef enum rpmTag_e {
     RPMTAG_FIRSTFREE_TAG	/*!< internal */
 } rpmTag;
 
-#define RPMTAG_NOT_FOUND		-1
+#define RPMTAG_NOT_FOUND		((uint32_t)-1)
 #define	RPMTAG_EXTERNAL_TAG		1000000
 
 /** \ingroup rpmtag
